Add decode top-k split path coverage
